# apps/js/docs/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "BeliMudah API",
		"version": "1.0",
		"description": "E-commerce API for the BeliMudah storefront.",
		"contact": {
			"name": "Ryan Suhartanto",
			"url": "https://github.com/ryansuhartanto/koda-b8-backend",
			"email": "suhartanto@kekkon.nexus"
		},
		"license": {
			"name": "MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3002"
		}
	],
	"components": {
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		},
		"schemas": {
			"Problem": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer"
					},
					"detail": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"status"
				]
			},
			"VariantOption": {
				"type": "object",
				"properties": {
					"option": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"option",
					"value"
				]
			},
			"ProductVariant": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"sku": {
						"type": "string"
					},
					"stock": {
						"type": "integer"
					},
					"price_idr": {
						"type": "integer"
					},
					"original_price_idr": {
						"type": "integer"
					},
					"options": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/VariantOption"
						},
						"uniqueItems": false
					}
				},
				"required": [
					"id",
					"options",
					"original_price_idr",
					"price_idr",
					"stock"
				]
			},
			"Product": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"brand": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"urls": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"uniqueItems": false
					},
					"price_idr": {
						"type": "integer"
					},
					"original_price_idr": {
						"type": "integer"
					},
					"stock": {
						"type": "integer"
					},
					"rating": {
						"type": "number"
					},
					"rating_count": {
						"type": "integer"
					},
					"variants": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductVariant"
						},
						"uniqueItems": false
					}
				},
				"required": [
					"created_at",
					"id",
					"name",
					"rating_count",
					"stock",
					"updated_at"
				]
			},
...
		}
	},
	"paths": {
...
		"/products": {
			"get": {
				"summary": "List products",
				"tags": [
					"products"
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"description": "Match against the product name",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"description": "Category name",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "brand",
						"description": "Brand name",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"description": "One of newest, price_asc, price_desc, rating",
						"schema": {
							"type": "string",
							"enum": [
								"newest",
								"price_asc",
								"price_desc",
								"rating"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Rows to return, 1 to 100",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"in": "query",
						"name": "offset",
						"description": "Rows to skip",
						"schema": {
							"type": "integer",
							"default": 0
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {
							"Link": {
								"description": "RFC 8288 pagination links: self, first, last, prev, next",
								"schema": {
									"type": "string"
								}
							},
							"X-Total-Count": {
								"description": "Rows matching the filter, ignoring limit and offset",
								"schema": {
									"type": "integer"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid query",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
		"/products/{id_product}": {
			"get": {
				"summary": "Fetch one product and its variants",
				"tags": [
					"products"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id_product",
						"required": true,
						"description": "Product sqid",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					},
					"404": {
						"description": "No such product",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
...
	}
}